refactor(checker): split type-parameter-target elaboration into its own module
Keep error_reporter/assignability.rs under the 2000-LOC arch-guard limit by
moving unrelated_type_parameter_target_related_info into a dedicated
assignability_type_parameter_target module (no behavior change).
