+++ crates/tsz-checker/src/error_reporter/mod.rs
@@ -24,6 +24,7 @@ mod assignability_normalized_union;
 mod assignability_numeric_display;
 mod assignability_satisfies;
 mod assignability_type_helpers;
+mod assignability_type_parameter_target;
 mod call_errors;
 mod call_errors_anchors;
 mod conditional_alias_display;
